+++ views/pages/nosqlInjection.gohtml
@@ -8,7 +8,7 @@
     <h4 class="sub-header">
       <code>{{.Name}}</code>
     </h4>
-    <form method="{{.Method}}" action="{{.URL}}/unsafe" target="_blank">
+    <form method="{{index $routeInfo.Inputs 0 | methodFromInput}}" action="{{.URL}}/unsafe" target="_blank">
       <div class="form-group">
         <label>Input</label>
         <input name="input" class="form-control" value='" || "1" == "1' />
