+++ views/pages/unvalidatedRedirect.gohtml
@@ -6,7 +6,7 @@
 <div class="row" style="min-height: 200px;">
   <div class="col-xs-12 col-sm-6" style="padding-bottom: 30px;">
     <h4 class="sub-header"><code>{{.Name}}</code></h4>
-    <form method="{{.Method}}" action="{{.URL}}unsafe" target="_blank">
+    <form method="{{index $routeInfo.Inputs 0 | methodFromInput}}" action="{{.URL}}unsafe" target="_blank">
       <div class="form-group">
         <label>URL</label>
         <input name="input" class="form-control" value="http://www.example.com" />
