+++ views/pages/xxe.gohtml
@@ -14,7 +14,7 @@
     <h4 class="sub-header">
       <code>{{.Name}}</code>
     </h4>
-    <form method="{{.Method}}" action="{{.URL}}/unsafe" target="_blank">
+    <form method="{{index $routeInfo.Inputs 0 | methodFromInput}}" action="{{.URL}}/unsafe" target="_blank">
       <div class="form-group">
         <input name="input" class="form-control"
           value="&lt;!DOCTYPE read-fs [&lt;!ELEMENT read-fs ANY &gt;&lt;!ENTITY passwd SYSTEM &quot;file:///etc/passwd&quot; &gt;]&gt;&lt;users&gt;&lt;user&gt;&lt;read-fs&gt;&amp;passwd;&lt;/read-fs&gt;&lt;name&gt;C.K Frode&lt;/name&gt;&lt;/user&gt;&lt;/users&gt;" />
